+++ cheatGUI/src/hacks/minigame.ts
@@ -1,7 +1,8 @@
 // Minigame Hacks
 
 // BEGIN IMPORTS
-import { category, Toggler } from "../index"; // Import the Cheat GUI bases.
+import { category } from "../index"; // Import the Cheat GUI bases.
+import Toggler from "../class/Toggler";
 import { _ } from "../utils/util";  // Import Prodigy Typings.
 import { Toast } from "../utils/swal"; // Import Toast and NumberInput from swal
 // END IMPORTS
