+++ cheatGUI/src/index.ts
@@ -3,7 +3,7 @@
 
 import { io } from "socket.io-client"; // Import socket.io-client
 import "./style.scss"; // Import SCSS style
-import { saveCharacter, _ } from "./utils/util"; // Import Prodigy typings
+import { _ } from "./utils/util"; // Import Prodigy typings
 import { statusMessage } from "./utils/status"; // Import status message
 import Swal from "sweetalert2"; // Import Swal
 import { License, NoLicense } from "./utils/swal";
@@ -92,92 +92,6 @@ subtitle.style.color = "white";
 menuleft.append(subtitle);
 
 
-
-export class Hack {
-	public element: HTMLButtonElement;
-	public name: String;
-	private description: String;
-
-	constructor (
-		public parent: HTMLDivElement,
-		name?: string,
-		description?: string
-	) {
-		this.name = "";
-		this.description = "";
-		this.element = document.createElement("button");
-		this.element.classList.add("menu-hack");
-		this.parent.append(this.element);
-
-		if (name) this.setName(name);
-		if (description) this.setDesc(description);
-	}
-
-	setName (name: string) {
-		this.element.innerText = name;
-		this.name = name;
-		return this;
-	}
-
-	setClick (event: () => unknown) {
-		this.element.onclick = async () => {
-			await event();
-			saveCharacter();
-			console.log(`Triggered ${this.name}.`);
-		};
-		return this;
-	}
-
-	setDesc (desc: string) {
-		this.element.title = desc;
-		this.description = desc;
-		return this;
-	}
-}
-
-export class Toggler extends Hack {
-	enabled?: () => unknown;
-	disabled?: () => unknown;
-	constructor (
-		public parent: HTMLDivElement,
-		name?: string,
-		description?: string
-	) {
-		super(parent, name, description);
-		this.element.setAttribute("status", "false");
-		this.setClick(async () => {
-			this.status = !this.status;
-			if (this.status) {
-				localStorage.setItem(this.name, "true");
-				await this.enabled?.();
-			} else {
-				localStorage.setItem(this.name, "false");
-				await this.disabled?.();
-			}
-		});
-	}
-
-	get status () {
-		return JSON.parse(this.element.getAttribute("status")!) as boolean;
-	}
-
-	set status (val) {
-		this.element.setAttribute("status", val.toString());
-	}
-
-	setEnabled (event: () => unknown) {
-		this.enabled = event;
-		if (localStorage.getItem(this.name) === "true") {
-			this.element.click();
-		}
-		return this;
-	}
-
-	setDisabled (event: () => unknown) {
-		this.disabled = event;
-		return this;
-	}
-}
 
 export const category = {
 	player: addArea("Player Hacks"),
