+++ cheatGUI/src/utils/chat.ts
@@ -1,0 +1,3 @@
+export default function addChat () {
+    
+}
